Fix CRLF on Windows (hopefully)
I don't have a Windows machine to test this. This uses .gitattributes to
hopefully force git to *not* turn LF into CRLF for the one file in which it
currently matters. In the future, a more global setting may be in order.

An option that DOESN'T work: using RUN in the Dockerfile. This leads to the
command not having the environment variable from docker-compose. i'm sure there
are workarounds for this, but i didn't want to chase them