Add basic client route for creating referrals
Also links said referrals from a job post

Note that the *client-side route* contains an ID so that the exact link to send
a referral to a position can be copied and pasted to a friend.

Also note that CreateReferral.tsx is heavily copy-pasted from
Listing.tsx. There are a couple of things there that really should be
de-duplicated, styles in particular.